# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * MilestoneVault API — off-chain metadata for on-chain escrow agreements
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Agreement {
  id: number;
  contractAddress: string;
  investorAddress: string;
  builderAddress: string;
  projectName: string;
  description: string;
  /** @nullable */
  txHash?: string | null;
  chainId: number;
  createdAt: string;
}

export interface AgreementInput {
  contractAddress: string;
  investorAddress: string;
  builderAddress: string;
  projectName: string;
  description: string;
  txHash?: string;
  chainId: number;
}

export type RequestMetaRequestType = typeof RequestMetaRequestType[keyof typeof RequestMetaRequestType];


export const RequestMetaRequestType = {
  milestone: 'milestone',
  adhoc: 'adhoc',
} as const;

export interface RequestMeta {
  id: number;
  agreementAddress: string;
  onchainRequestId: number;
  requestType: RequestMetaRequestType;
  /** @nullable */
  milestoneIndex?: number | null;
  /** @nullable */
  reason?: string | null;
  /** @nullable */
  evidenceUrl?: string | null;
  createdAt: string;
}

export type RequestMetaInputRequestType = typeof RequestMetaInputRequestType[keyof typeof RequestMetaInputRequestType];


export const RequestMetaInputRequestType = {
  milestone: 'milestone',
  adhoc: 'adhoc',
} as const;

export interface RequestMetaInput {
  onchainRequestId: number;
  requestType: RequestMetaInputRequestType;
  /** @nullable */
  milestoneIndex?: number | null;
  reason?: string;
  evidenceUrl?: string;
}

export type ListAgreementsParams = {
/**
 * Filter by investor address
 */
investor?: string;
/**
 * Filter by builder address
 */
builder?: string;
};