# lib/api-zod/src/generated/types/agreement.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * MilestoneVault API — off-chain metadata for on-chain escrow agreements
 * OpenAPI spec version: 0.1.0
 */

export interface Agreement {
  id: number;
  contractAddress: string;
  investorAddress: string;
  builderAddress: string;
  projectName: string;
  description: string;
  /** @nullable */
  txHash?: string | null;
  chainId: number;
  createdAt: string;
}