# lib/api-zod/src/generated/types/requestMeta.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * MilestoneVault API — off-chain metadata for on-chain escrow agreements
 * OpenAPI spec version: 0.1.0
 */
import type { RequestMetaRequestType } from './requestMetaRequestType';

export interface RequestMeta {
  id: number;
  agreementAddress: string;
  onchainRequestId: number;
  requestType: RequestMetaRequestType;
  /** @nullable */
  milestoneIndex?: number | null;
  /** @nullable */
  reason?: string | null;
  /** @nullable */
  evidenceUrl?: string | null;
  createdAt: string;
}