# lib/api-zod/src/generated/types/requestMetaInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * MilestoneVault API — off-chain metadata for on-chain escrow agreements
 * OpenAPI spec version: 0.1.0
 */
import type { RequestMetaInputRequestType } from './requestMetaInputRequestType';

export interface RequestMetaInput {
  onchainRequestId: number;
  requestType: RequestMetaInputRequestType;
  /** @nullable */
  milestoneIndex?: number | null;
  reason?: string;
  evidenceUrl?: string;
}